requirements.txt
affine==2.4.0
altgraph==0.17.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.6.15
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fiona==1.10.1
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pefile==2023.2.7
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==27.0.0
rasterio==1.4.3
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
wcwidth==0.2.13
